Move some util functions to common folder
