+++ src/renderer/containers/actions/ActionSaveReplayBuffer.tsx
@@ -4,7 +4,8 @@ import { ActionTypeGenerator, Context } from "@vinceau/event-actions";
 import { produce } from "immer";
 
 import { DelayInput, NotifyInput } from "@/components/InlineInputs";
-import { delay as waitMillis, notify as sendNotification, parseSecondsDelayValue } from "@/lib/utils";
+import { notify as sendNotification } from "@/lib/utils";
+import { delay as waitMillis, parseSecondsDelayValue } from "common/utils";
 import { ActionComponent } from "./types";
 
 import { CustomIcon } from "@/components/CustomIcon";
