+++ src/renderer/lib/realtime.ts
@@ -5,8 +5,9 @@ import { withLatestFrom } from "rxjs/operators";
 import { dispatcher } from "@/store";
 import { exampleComboType, exampleDeathStockType, exampleGameEnd, exampleGameStart, exampleSpawnStockType,
         generateComboContext, generateGameEndContext, generateGameStartContext, generateGlobalContext, generateStockContext } from "common/context";
+import { isDevelopment } from "common/utils";
 import { eventActionManager } from "../containers/actions";
-import { isDevelopment, notify } from "./utils";
+import { notify } from "./utils";
 
 export enum ActionEvent {
     GAME_START = "game-start",
