+++ src/renderer/lib/utils.ts
@@ -3,15 +3,12 @@ import * as url from "url";
 
 import { DolphinRecorderOptions, openComboInDolphin } from "@/lib/dolphin";
 import { Message } from "common/types";
-import { remote, shell } from "electron";
+import { remote } from "electron";
 import fs from "fs-extra";
 import { ipc } from "./rendererIpc";
 
 import { DolphinQueueFormat } from "@vinceau/slp-realtime";
-
-export const delay = async (ms: number): Promise<void> => {
-    await new Promise(resolve => setTimeout(resolve, ms));
-};
+import { isDevelopment } from "common/utils";
 
 const folderOptions = {
     properties: ["openDirectory"],
@@ -59,10 +56,6 @@ export const notify = (message: string, title?: string) => {
     );
 };
 
-export const isDevelopment = process.env.NODE_ENV !== "production";
-
-export const isMacOrWindows = process.platform === "darwin" || process.platform === "win32";
-
 // see https://github.com/electron-userland/electron-webpack/issues/99#issuecomment-459251702
 export const getStatic = (val: string): string => {
     if (isDevelopment) {
@@ -90,46 +83,4 @@ export const loadDolphinQueue = async (): Promise<DolphinQueueFormat | null> => 
         return fs.readJson(p[0]);
     }
     return null;
-};
-
-/**
- * Open the specified file in the system file explorer.
- * If the file doesn't exist, open the parent directory.
- * If the parent directory doesn't exist, try the parent's parent directory etc.
- */
-export const openFileOrParentFolder = (fileName: string) => {
-    let opened = shell.showItemInFolder(fileName);
-    let parentFolder = fileName;
-    while (!opened) {
-        parentFolder = path.dirname(parentFolder);
-        opened = shell.openItem(parentFolder);
-    }
-};
-
-export const parseSecondsDelayValue = (defaultSeconds: number, delaySeconds?: string): number => {
-    let seconds = parseInt(delaySeconds || defaultSeconds.toString(), 10);
-    if (isNaN(seconds)) {
-        seconds = defaultSeconds;
-    }
-    return seconds;
-};
-
-export const capitalize = (s: string) => {
-    return s.charAt(0).toUpperCase() + s.slice(1);
-};
-
-export const framesToMillis = (frames: number): number => {
-    return framesToSeconds(frames) * 1000;
-};
-
-export const framesToSeconds = (frames: number): number => {
-    return frames / 60.0;
-};
-
-export const secondsToFrames = (secs: number): number => {
-    return secs * 60.0;
-};
-
-export const millisToFrames = (ms: number): number => {
-    return secondsToFrames(ms / 1000);
 };
